granularity of time logging changed to Double from Long

+++ examples/src/main/java/com/asgow/ciel/examples/mapreduce/common/MapTask.java
@@ -72,14 +72,14 @@ public class MapTask implements ConstantNumOutputsTask {
 	        // call map logic
 	        run(bufferedReader, tempDos, nReducers);
 	        System.out.println("MapReduce: Map " + Integer.toString(id) + " logic completed in "
-					 + Long.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
+					 + Double.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
 			// now sort the temp files with 50 Mb mem limit
 			TextFileSorter sorter = new TextFileSorter(new SortConfig().withMaxMemoryUsage(50 * 1000 * 1000));
 			for(int i = 0; i < nReducers; i++) {
 				sorter.sort(new FileInputStream(tempFiles[i]), outputs[i]);
 			}
 			System.out.println("MapReduce: Map " + Integer.toString(id) + " sort completed in "
-					 + Long.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
+					 + Double.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
         } catch (Exception e) {
         	 System.out.println("MapReduce: Exception while running MapTask" + " for job: " + jobID);
         	 e.printStackTrace();
@@ -97,7 +97,7 @@ public class MapTask implements ConstantNumOutputsTask {
         }
 
         System.out.println("MapReduce: Map " + Integer.toString(id) + " finished in "
-		 + Long.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
+		 + Double.toString((System.currentTimeMillis() - startTime)/1000) + " secs at " + dateTime.getCurrentDateTime() + " for job: " + jobID);
 	}
 
 	public void setup() {
